Synced sqlite3utils with private and fixed a small recursive bug in SIPEngine that was already fixed in private
git-svn-id: http://wush.net/svn/range/software/public/openbts/trunk@4912 19bc5d8c-e614-43d4-8b26-e1612bc8e597

+++ SIP/SIPEngine.cpp
@@ -114,7 +114,7 @@ ostream& SIP::operator<<(ostream& os, SIPState s)
 {
 	const char* str = SIPStateString(s);
 	if (str) os << str;
-	else os << "?" << s << "?";
+	else os << "?" << ((int)s) << "?";
 	return os;
 }
